fix: remove the race condition once for all by merging the start/stop test into the lxc tests
